General bug fixes (#142)
* Core/Bitset: Fix TestAll method

* Fix documentation

* Fix color and their conversions

* Core/ByteStream: Fix return of Write

* Fix compiler warnings

* Math/Algorithm: Fix angle normalization

* Math/BoundingVolume: Fix lerp

* Math: Fix relation between Matrix4 and Quaternion

* More tests

* X11/Window: Fix mouse moved event generated when doing Mouse::SetPosition

* Update ChangeLog

* Should fix compilation on Windows

* Should fix compilation on Windows

Forgot to include array for Windows

+++ include/Nazara/Core/String.inl
@@ -121,7 +121,7 @@ namespace std
 				const char* ptr = str.GetConstBuffer();
 
 				do
-					h = ((h << 5) + h) + *ptr;
+					h = ((h << 5) + h) + static_cast<size_t>(*ptr);
 				while (*++ptr);
 			}
